Add fnl build target
Using gcc 4.9 and glibc.

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1348493002 .

+++ services/http_server/http_request_parser.cc
@@ -139,10 +139,10 @@ HttpRequestParser::ParseResult HttpRequestParser::ParseHeaders() {
 
   // If we ever want to support really large content length (currently pipe max
   // 256 MB), then we'll have to stream data from parser to handler.
-  MojoCreateDataPipeOptions options = {sizeof(MojoCreateDataPipeOptions),
-                                       MOJO_CREATE_DATA_PIPE_OPTIONS_FLAG_NONE,
-                                       1,
-                                       declared_content_length};
+  MojoCreateDataPipeOptions options = {
+      sizeof(MojoCreateDataPipeOptions),
+      MOJO_CREATE_DATA_PIPE_OPTIONS_FLAG_NONE, 1,
+      static_cast<uint32_t>(declared_content_length)};
   MojoResult result = CreateDataPipe(
       &options, &producer_handle_, &http_request_->body);
   if (result != MOJO_RESULT_OK) {
